requirements.txt
anytree
appdirs
aria2p
attrdict
beautifulsoup4
cfscrape
feedparser
flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
lk21
lxml
megasdkrestclient
pillow
psutil
psycopg2-binary
pybase64
pyrogram
python-dotenv
python-magic
python-telegram-bot
qbittorrent-api
requests
speedtest-cli
six
telegraph
tenacity
tgCrypto
urllib3
yt-dlp
